web/crypt0r-part-2/server-files/requirements.txt
Flask==0.12.2
PyYAML==3.12

